banking_spark_project/local_requirements.txt
pyspark==3.3.0
apache-airflow==2.7.1
pandas
openpyxl
sqlalchemy
psycopg2-binary